9C-1141 Collections Processes
We create in the `processes` module a special package for the
collection processes, messages, converters, and tests.
We change the necessary imports in the `api` module.
