+++ modules/processes/src/test/scala/cards/nine/processes/TestData.scala
@@ -2,8 +2,7 @@ package cards.nine.processes
 
 import java.sql.Timestamp
 import java.time.Instant
-
-import cards.nine.commons.NineCardsErrors.{ CountryNotFound, RankingNotFound, SharedCollectionNotFound }
+import cards.nine.commons.NineCardsErrors.{ CountryNotFound, RankingNotFound }
 import cards.nine.commons.NineCardsService
 import cards.nine.domain.account.{ AndroidId, DeviceToken }
 import cards.nine.domain.analytics._
@@ -12,11 +11,8 @@ import cards.nine.domain.market.{ Localization, MarketCredentials, MarketToken }
 import cards.nine.domain.pagination.Page
 import cards.nine.processes.NineCardsServices.NineCardsServices
 import cards.nine.processes.converters.Converters
-import cards.nine.processes.messages.SharedCollectionMessages._
-import cards.nine.services.free.domain.Firebase.{ NotificationIndividualResult, SendNotificationResponse }
 import cards.nine.services.free.domain.Ranking.GoogleAnalyticsRanking
-import cards.nine.services.free.domain.{ SharedCollection ⇒ SharedCollectionServices, _ }
-import cards.nine.services.free.interpreter.collection.Services.{ SharedCollectionData ⇒ SharedCollectionDataServices }
+import cards.nine.services.free.domain.{ Country, Installation }
 import org.joda.time.{ DateTime, DateTimeZone }
 
 object TestData {
@@ -32,8 +28,6 @@ object TestData {
   val canonicalId = 0
 
   val category = "SOCIAL"
-
-  val collectionId = 1l
 
   val community = true
 
@@ -88,8 +82,6 @@ object TestData {
   val success = 1
 
   val token = "6d54dfed-bbcf-47a5-b8f2-d86cf3320631"
-
-  val updatedCollectionsCount = 1
 
   val userId = Option(publisherId)
 
@@ -150,25 +142,6 @@ object TestData {
       token        = MarketToken(token)
     )
 
-    val collection = SharedCollectionServices(
-      id               = collectionId,
-      publicIdentifier = publicIdentifier,
-      userId           = userId,
-      publishedOn      = publishedOnTimestamp,
-      author           = author,
-      name             = name,
-      views            = views,
-      category         = category,
-      icon             = icon,
-      community        = community,
-      packages         = packagesName map (_.value)
-    )
-
-    val collectionWithSubscriptions = SharedCollectionWithAggregatedInfo(
-      sharedCollectionData = collection,
-      subscriptionsCount   = subscriptionsCount
-    )
-
     val country = Country(
       isoCode2  = countryIsoCode2,
       isoCode3  = Option(countryIsoCode3),
@@ -177,12 +150,6 @@ object TestData {
     )
 
     val countryNotFoundError = CountryNotFound(s"Country with ISO code2 US doesn't exist")
-
-    val sharedCollectionNotFoundError = SharedCollectionNotFound("Shared collection not found")
-
-    val createPackagesStats = PackagesStats(packagesName.size, None)
-
-    val updatePackagesStats = PackagesStats(addedPackagesCount, Option(removedPackagesCount))
 
     val appInfoList = packagesName map { packageName ⇒
       FullCard(packageName, "Germany", List(appCategory), "100.000+", true, icon, Nil, stars)
@@ -193,116 +160,6 @@ object TestData {
       userId      = subscriberId,
       deviceToken = Option(deviceToken),
       androidId   = androidId
-    )
-
-    val sendNotificationResponse = SendNotificationResponse(
-      multicastIds = List(multicastId),
-      success      = success,
-      failure      = failure,
-      canonicalIds = canonicalId,
-      results      = List(NotificationIndividualResult(
-        message_id      = Option(messageId),
-        registration_id = None,
-        error           = None
-      ))
-    )
-
-    val sharedCollectionDataServices = SharedCollectionDataServices(
-      publicIdentifier = publicIdentifier,
-      userId           = userId,
-      publishedOn      = publishedOnTimestamp,
-      author           = author,
-      name             = name,
-      views            = views,
-      category         = category,
-      icon             = icon,
-      community        = community,
-      packages         = packagesName map (_.value)
-    )
-
-    val sharedCollectionData = SharedCollectionData(
-      publicIdentifier = publicIdentifier,
-      userId           = userId,
-      publishedOn      = publishedOnDatetime,
-      author           = author,
-      name             = name,
-      views            = Option(views),
-      category         = category,
-      icon             = icon,
-      community        = community,
-      packages         = packagesName
-    )
-
-    val sharedCollection = SharedCollection(
-      publicIdentifier = publicIdentifier,
-      publishedOn      = new DateTime(publishedOnTimestamp.getTime),
-      author           = author,
-      name             = name,
-      views            = views,
-      category         = category,
-      icon             = icon,
-      community        = community,
-      owned            = true,
-      packages         = packagesName
-    )
-
-    val sharedCollectionWithSubscriptions = SharedCollection(
-      publicIdentifier   = publicIdentifier,
-      publishedOn        = new DateTime(publishedOnTimestamp.getTime),
-      author             = author,
-      name               = name,
-      views              = views,
-      category           = category,
-      icon               = icon,
-      community          = community,
-      owned              = true,
-      packages           = packagesName,
-      subscriptionsCount = Option(subscriptionsCount)
-    )
-
-    val sharedCollectionUpdateInfo = SharedCollectionUpdateInfo(
-      title = name
-    )
-
-    val sharedCollectionWithAppsInfo = SharedCollectionWithAppsInfo(
-      collection = sharedCollection,
-      appsInfo   = appInfoList
-    )
-
-    val sharedCollectionWithAppsInfoBasic = SharedCollectionWithAppsInfo(
-      collection = sharedCollection,
-      appsInfo   = appInfoList map (_.toBasic)
-    )
-
-    val sharedCollectionWithAppsInfoAndSubscriptions = SharedCollectionWithAppsInfo(
-      collection = sharedCollectionWithSubscriptions,
-      appsInfo   = appInfoList map (_.toBasic)
-    )
-
-    val subscription = SharedCollectionSubscription(
-      sharedCollectionId       = collectionId,
-      userId                   = subscriberId,
-      sharedCollectionPublicId = publicIdentifier
-    )
-
-    val updatedSubscriptionsCount = 1
-  }
-
-  object Messages {
-
-    import Values._
-
-    val createCollectionRequest: CreateCollectionRequest = CreateCollectionRequest(
-      collection = sharedCollectionData
-    )
-
-    val createCollectionResponse = CreateOrUpdateCollectionResponse(
-      publicIdentifier = publicIdentifier,
-      packagesStats    = createPackagesStats
-    )
-
-    val getCollectionByPublicIdentifierResponse = GetCollectionByPublicIdentifierResponse(
-      data = sharedCollectionWithAppsInfo
     )
 
   }
